Refactor + clippy
